CIRC-1358 Move getFirstAvailableItemByInstanceId method to ItemRepository
